menyelesaikan fitur buat produk

+++ controllers/Forum/post.js
@@ -146,7 +146,17 @@ const getOnePost = async (req, res) => {
 
 const deletePost = async (req, res) => {
     const id = parseInt(req.params.id);
+    const userID = req.user.id;
     try {
+        const commentIduser = await Comment.findByPk(id);
+        
+        if (!commentIduser) {
+            return res.status(404).json({ message: "Komentar tidak ditemukan." });
+        }
+
+        if (commentIduser.userId !== userID) {
+            res.status(500).json({ message: "maaf kamu tidak bisa mengedit komen" });
+        }
         await Post.destroy({where: {id:id} })
         res.status(200).json({ message: "Delete successful" });
     } catch (error) {
